requirements.txt
agno
openai
mcp
python-dotenv
requests

